requirements.txt
# Core dependencies
requests>=2.31.0
beautifulsoup4>=4.12.2
pandas>=2.0.0

# Google APIs
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
gspread>=5.10.0
gspread-formatting>=1.1.2

# Configuration management
python-dotenv>=1.0.0



